requirements.txt
google-genai
python-dotenv
zhipuai
socksio
